For #65. Allows quick deletion only if the group owner is the only member of the group, otherwise navigates to a separate delete group page.

+++ mod/groups/actions/groups/delete.php
@@ -16,27 +16,41 @@ if (!$entity->canEdit()) {
 	forward(REFERER);
 }
 
-if (($entity) && ($entity instanceof ElggGroup)) {
-	// delete group icons
-	$owner_guid = $entity->owner_guid;
-	$prefix = "groups/" . $entity->guid;
-	$imagenames = array('.jpg', 'tiny.jpg', 'small.jpg', 'medium.jpg', 'large.jpg');
-	$img = new ElggFile();
-	$img->owner_guid = $owner_guid;
-	foreach ($imagenames as $name) {
-		$img->setFilename($prefix . $name);
-		$img->delete();
-	}
+// check member count
+$count = elgg_get_entities_from_relationship(array(
+		'relationship' => 'member',
+		'relationship_guid' => $guid,
+		'inverse_relationship' => true,
+		'type' => 'user',
+		'limit' => 20,
+));
+if ( count( $count ) == 1 ){
 
-	// delete group
-	if ($entity->delete()) {
-		system_message(elgg_echo('group:deleted'));
+	if (($entity) && ($entity instanceof ElggGroup)) {
+		// delete group icons
+		$owner_guid = $entity->owner_guid;
+		$prefix = "groups/" . $entity->guid;
+		$imagenames = array('.jpg', 'tiny.jpg', 'small.jpg', 'medium.jpg', 'large.jpg');
+		$img = new ElggFile();
+		$img->owner_guid = $owner_guid;
+		foreach ($imagenames as $name) {
+			$img->setFilename($prefix . $name);
+			$img->delete();
+		}
+
+		// delete group
+		if ($entity->delete()) {
+			system_message(elgg_echo('group:deleted'));
+		} else {
+			register_error(elgg_echo('group:notdeleted'));
+		}
 	} else {
 		register_error(elgg_echo('group:notdeleted'));
 	}
-} else {
-	register_error(elgg_echo('group:notdeleted'));
+
+	$url_name = elgg_get_logged_in_user_entity()->username;
+	forward(elgg_get_site_url() . "groups/member/{$url_name}");
 }
 
-$url_name = elgg_get_logged_in_user_entity()->username;
-forward(elgg_get_site_url() . "groups/member/{$url_name}");
+else
+	forward(elgg_get_site_url() . "groups/delete/{$guid}");
